+++ conf/config.properties
@@ -4,7 +4,7 @@
 
 ################### 被测数据库配置 #######################
 # 被测试的数据库，目前的格式为{name}{-version}{-insert mode}(注意-号)
-# name：IoTDB, InfluxDB, OpenTSDB, CTSDB, KairosDB, TimescaleDB, FakeDB, TaosDB
+# name：IoTDB, InfluxDB, OpenTSDB, KairosDB, TimescaleDB, FakeDB, TaosDB
 # version：以IoTDB为例，有09、010、011、012
 # insert mode：以IoTDB为例，有JDBC、SESSION_BY_TABLET、SESSION_BY_RECORD、SESSION_BY_RECORDS、SESSION_BY_POOL
 DB_SWITCH=IoTDB-012-SESSION_BY_TABLET
