+++ core/pom.xml
@@ -32,19 +32,19 @@
 
   <dependencies>
     <dependency>
-      <groupId>org.kairosdb</groupId>
-      <artifactId>client</artifactId>
-      <version>3.0.0</version>
-    </dependency>
-    <dependency>
       <groupId>commons-cli</groupId>
       <artifactId>commons-cli</artifactId>
       <version>${common.cli.version}</version>
     </dependency>
     <dependency>
-      <groupId>org.influxdb</groupId>
-      <artifactId>influxdb-java</artifactId>
-      <version>2.7</version>
+      <groupId>org.apache.commons</groupId>
+      <artifactId>commons-lang3</artifactId>
+      <version>3.7</version>
+    </dependency>
+    <dependency>
+      <groupId>org.weakref</groupId>
+      <artifactId>jmxutils</artifactId>
+      <version>1.21</version>
     </dependency>
     <dependency>
       <groupId>postgresql</groupId>
@@ -75,11 +75,6 @@
       <groupId>com.clearspring.analytics</groupId>
       <artifactId>stream</artifactId>
       <version>2.9.5</version>
-    </dependency>
-    <dependency>
-      <groupId>com.taosdata.jdbc</groupId>
-      <artifactId>taos-jdbcdriver</artifactId>
-      <version>2.0.12</version>
     </dependency>
     <dependency>
       <groupId>org.apache.kafka</groupId>
