+++ core/src/main/java/cn/edu/tsinghua/iotdb/benchmark/conf/Constants.java
@@ -132,10 +132,10 @@ public class Constants {
     public static final String IOTDB010_SESSION_CLASS = "cn.edu.tsinghua.iotdb.benchmark.iotdb010.IoTDBSession";
     public static final String IOTDB009_JDBC_CLASS = "cn.edu.tsinghua.iotdb.benchmark.iotdb009.IoTDB";
     public static final String IOTDB009_SESSION_CLASS = "cn.edu.tsinghua.iotdb.benchmark.iotdb009.IoTDBSession";
-    public static final String INFLUXDB_CLASS = "cn.edu.tsinghua.iotdb.benchmark.tsdb.influxdb.InfluxDB";
+    public static final String INFLUXDB_CLASS = "cn.edu.tsinghua.iotdb.benchmark.influxdb.InfluxDB";
     public static final String FAKEDB_CLASS = "cn.edu.tsinghua.iotdb.benchmark.tsdb.fakedb.FakeDB";
-    public static final String KAIROSDB_CLASS = "cn.edu.tsinghua.iotdb.benchmark.tsdb.kairosdb.KairosDB";
-    public static final String OPENTSDB_CLASS = "cn.edu.tsinghua.iotdb.benchmark.tsdb.opentsdb.OpenTSDB";
-    public static final String TIMESCALEDB_CLASS = "cn.edu.tsinghua.iotdb.benchmark.tsdb.timescaledb.TimescaleDB";
-    public static final String TAOSDB_CLASS = "cn.edu.tsinghua.iotdb.benchmark.tsdb.taosdb.TaosDB";
+    public static final String KAIROSDB_CLASS = "cn.edu.tsinghua.iotdb.benchmark.kairosdb.KairosDB";
+    public static final String OPENTSDB_CLASS = "cn.edu.tsinghua.iotdb.benchmark.opentsdb.OpenTSDB";
+    public static final String TIMESCALEDB_CLASS = "cn.edu.tsinghua.iotdb.benchmark.timescaledb.TimescaleDB";
+    public static final String TAOSDB_CLASS = "cn.edu.tsinghua.iotdb.benchmark.taosdb.TaosDB";
 }
