+++ influxdb/README.md
@@ -1,0 +1,6 @@
+Benchmark influxdb
+---
+This project is using iotdb-benchmark to test influxdb
+
+# environment
+1. influxdb: 1.4
