+++ iotdb-0.12/src/main/test/cn/edu/tsinghua/iotdb/benchmark/TestEntrance.java
