+++ pom.xml
@@ -11,7 +11,11 @@
   <url>http://maven.apache.org</url>
   <modules>
     <module>core</module>
-    <!--<module>iotdb-0.10</module>-->
+    <module>influxdb</module>
+    <module>kairosdb</module>
+    <module>opentsdb</module>
+    <module>taosdb</module>
+    <module>timescaledb</module>
     <module>iotdb-0.12</module>
     <module>iotdb-0.11</module>
     <module>iotdb-0.10</module>
