+++ timescaledb/src/main/test/cn/edu/tsinghua/iotdb/benchmark/TestEntrance.java
@@ -1,0 +1,9 @@
+package cn.edu.tsinghua.iotdb.benchmark;
+
+import java.sql.SQLException;
+
+public class TestEntrance {
+    public static void main(String[] args) throws SQLException {
+        App.main(args);
+    }
+}
